Refactor gameplay.js to function for reusing, remove unused files
